Added folders regarding code
Added two folders: One folder is code written by Dr. Leiner as a tutorial in order to learn how to analyze the Coude Telescope data provided. One file is to analyze the spectra themselves, one is to learn how to use saphires to get the broadening function, and the last is to use the saphires and the broadening functions to calculate the radial velocities and the vsin(i).

The other folder, in it's current state, consists of one notebook to automate the radial velocity and vsin(i) of the SSGs. I will return to it in due time, but at the moment I am trying to figure out how to rerun the data to produce .p files so that we may have sufficient data to calculate the data for all the radial velocities.