Automated Broadening function calculations
This notebook needs to be tweaked a bit to be fully optimized. With that being said, this notebook takes .ls files from observations using the Coude telescope to calculate broadening functions that we save as.p files. This notebook is important as we are missing many .p files for many of the stars in the SSG sample.